Spoonful of generics tests.
