+++ independent-projects/arc/tests/src/test/java/io/quarkus/arc/test/injection/assignability/generics/Car.java
@@ -1,4 +1,4 @@
-package io.quarkus.arc.test.injection.generics;
+package io.quarkus.arc.test.injection.assignability.generics;
 
 import javax.enterprise.context.ApplicationScoped;
