+++ independent-projects/arc/tests/src/test/java/io/quarkus/arc/test/injection/assignability/generics/Engine.java
@@ -1,0 +1,4 @@
+package io.quarkus.arc.test.injection.assignability.generics;
+
+public interface Engine {
+}
