--- independent-projects/arc/tests/src/test/java/io/quarkus/arc/test/injection/generics/Engine.java
@@ -1,4 +1,0 @@
-package io.quarkus.arc.test.injection.generics;
-
-public interface Engine {
-}
